# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synapse2 REST API",
    "version": "0.5.4",
    "description": "Generated OpenAPI schema for Synapse2's REST surface. Loopback deployments have no HTTP auth; non-loopback deployments require SYNAPSE_MCP_TOKEN or OAuth bearer JWTs. REST actions require their action-specific scopes when auth is mounted."
  },
  "servers": [
    {
      "url": "http://localhost:40080",
      "description": "Default local development server"
    }
  ],
...
  "paths": {
...
    "/openapi.json": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "OpenAPI schema",
        "operationId": "getOpenApiSchema",
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "Generated OpenAPI schema for the REST surface",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/status": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Runtime status",
        "operationId": "getStatus",
        "security": [],
        "responses": {
          "200": {
            "description": "Runtime status with secrets redacted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/synapse2": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Dispatch a REST action",
        "description": "Thin REST shim over the shared service layer. MCP-only actions are not exposed here. Current REST actions: help, flux.docker.info, flux.docker.df, flux.docker.images, flux.docker.networks, flux.docker.volumes, flux.docker.pull, flux.docker.build, flux.docker.rmi, flux.docker.prune, flux.container.list, scout.nodes, scout.peek, scout.exec. When auth is mounted, each action requires its declared scope; synapse:write satisfies synapse:read.",
        "operationId": "dispatchSynapse2Action",
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              },
              "examples": {
                "help": {
                  "value": {
                    "action": "help",
                    "params": {}
                  }
                },
                "flux.docker.info": {
                  "value": {
                    "action": "flux.docker.info",
                    "params": {
                      "host": "myhost"
                    }
                  }
                },
                "flux.docker.df": {
                  "value": {
                    "action": "flux.docker.df",
                    "params": {}
                  }
                },
                "flux.docker.images": {
                  "value": {
                    "action": "flux.docker.images",
                    "params": {}
                  }
                },
                "flux.docker.networks": {
                  "value": {
                    "action": "flux.docker.networks",
                    "params": {}
                  }
                },
                "flux.docker.volumes": {
                  "value": {
                    "action": "flux.docker.volumes",
                    "params": {}
                  }
                },
                "flux.docker.pull": {
                  "value": {
                    "action": "flux.docker.pull",
                    "params": {
                      "host": "myhost",
                      "image": "nginx:latest"
                    }
                  }
                },
                "flux.docker.build": {
                  "value": {
                    "action": "flux.docker.build",
                    "params": {}
                  }
                },
                "flux.docker.rmi": {
                  "value": {
                    "action": "flux.docker.rmi",
                    "params": {}
                  }
                },
                "flux.docker.prune": {
                  "value": {
                    "action": "flux.docker.prune",
                    "params": {}
                  }
                },
                "flux.container.list": {
                  "value": {
                    "action": "flux.container.list",
                    "params": {
                      "state": "running"
                    }
                  }
                },
                "scout.nodes": {
                  "value": {
                    "action": "scout.nodes",
                    "params": {}
                  }
                },
                "scout.peek": {
                  "value": {
                    "action": "scout.peek",
                    "params": {
                      "host": "myhost",
                      "path": "/etc/hostname"
                    }
                  }
                },
                "scout.exec": {
                  "value": {
                    "action": "scout.exec",
                    "params": {
                      "host": "myhost",
                      "path": "/tmp",
                      "command": "hostname"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Action result. Shape depends on the requested action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Static bearer token in bearer mode; OAuth mode also uses bearer JWTs. Loopback mode does not require HTTP auth."
      }
    },
    "schemas": {
      "ActionName": {
        "type": "string",
        "enum": [
          "help",
          "flux.docker.info",
          "flux.docker.df",
          "flux.docker.images",
          "flux.docker.networks",
          "flux.docker.volumes",
          "flux.docker.pull",
          "flux.docker.build",
          "flux.docker.rmi",
          "flux.docker.prune",
          "flux.container.list",
          "scout.nodes",
          "scout.peek",
          "scout.exec"
        ],
        "description": "REST-capable action names served by POST /v1/synapse2."
      },
      "ActionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionName"
          },
          "params": {
            "type": "object",
            "description": "Action-specific parameters for the selected Synapse2 REST action.",
            "additionalProperties": true,
            "default": {}
          }
        }
      },
      "ActionResponse": {
        "type": "object",
        "description": "Action result. Shape depends on the requested action.",
        "additionalProperties": true
      },
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "server": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "transport": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "OverloadResponse": {
        "type": "object",
        "required": [
          "error",
          "retryable"
        ],
        "properties": {
          "error": {
            "type": "string",
            "const": "server overloaded"
          },
          "retryable": {
            "type": "boolean",
            "const": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated request lacks the required scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Concurrency limit reached; retry after the indicated delay",
        "headers": {
          "Retry-After": {
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OverloadResponse"
            }
          }
        }
      }
    }
  },
...
}